refactor: Further merge refactor
