+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/merge/MergeService.java
@@ -29,6 +29,7 @@ package org.hisp.dhis.merge;
 
 import javax.annotation.Nonnull;
 import org.hisp.dhis.common.IdentifiableObject;
+import org.hisp.dhis.feedback.ConflictException;
 import org.hisp.dhis.feedback.MergeReport;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -38,6 +39,20 @@ import org.springframework.transaction.annotation.Transactional;
  * @author david mackessy
  */
 public interface MergeService {
+
+  default MergeReport processMerge(@Nonnull MergeParams mergeParams) throws ConflictException {
+    MergeReport mergeReport = new MergeReport();
+
+    MergeRequest mergeRequest = validate(mergeParams, mergeReport);
+    if (mergeReport.hasErrorMessages())
+      throw new ConflictException("Merge validation error").setMergeReport(mergeReport);
+
+    MergeReport report = merge(mergeRequest, mergeReport);
+    if (report.hasErrorMessages())
+      throw new ConflictException("Merge error").setMergeReport(mergeReport);
+
+    return report;
+  }
 
   /**
    * This method transforms a {@link MergeParams} to a {@link MergeRequest}. If there are any
