+++ dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/controller/category/CategoryOptionController.java
@@ -49,7 +49,7 @@ import org.hisp.dhis.dxf2.webmessage.WebMessageUtils;
 import org.hisp.dhis.feedback.ConflictException;
 import org.hisp.dhis.feedback.MergeReport;
 import org.hisp.dhis.merge.MergeParams;
-import org.hisp.dhis.merge.MergeProcessor;
+import org.hisp.dhis.merge.MergeService;
 import org.hisp.dhis.security.RequiresAuthority;
 import org.hisp.dhis.webapi.controller.AbstractCrudController;
 import org.springframework.http.HttpStatus;
@@ -72,7 +72,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 @OpenApi.Document(classifiers = {"team:platform", "purpose:metadata"})
 public class CategoryOptionController extends AbstractCrudController<CategoryOption> {
   private final CategoryService categoryService;
-  private final MergeProcessor categoryOptionMergeProcessor;
+  private final MergeService categoryOptionMergeService;
 
   @ResponseBody
   @GetMapping(value = "orgUnits")
@@ -97,7 +97,7 @@ public class CategoryOptionController extends AbstractCrudController<CategoryOpt
 
     MergeReport report;
     try {
-      report = categoryOptionMergeProcessor.processMerge(params);
+      report = categoryOptionMergeService.processMerge(params);
     } catch (PersistenceException ex) {
       String helpfulMessage = getHelpfulMessage(ex);
       log.error("Error while processing CategoryOption merge: {}", helpfulMessage);
