+++ dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/controller/dataelement/DataElementController.java
@@ -42,7 +42,7 @@ import org.hisp.dhis.dxf2.webmessage.WebMessageUtils;
 import org.hisp.dhis.feedback.ConflictException;
 import org.hisp.dhis.feedback.MergeReport;
 import org.hisp.dhis.merge.MergeParams;
-import org.hisp.dhis.merge.MergeProcessor;
+import org.hisp.dhis.merge.MergeService;
 import org.hisp.dhis.security.RequiresAuthority;
 import org.hisp.dhis.webapi.controller.AbstractCrudController;
 import org.springframework.http.HttpStatus;
@@ -64,7 +64,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 public class DataElementController extends AbstractCrudController<DataElement> {
 
   private final DataElementService dataElementService;
-  private final MergeProcessor dataElementMergeProcessor;
+  private final MergeService dataElementMergeService;
 
   @ResponseStatus(HttpStatus.OK)
   @RequiresAuthority(anyOf = F_DATA_ELEMENT_MERGE)
@@ -75,7 +75,7 @@ public class DataElementController extends AbstractCrudController<DataElement> {
 
     MergeReport report;
     try {
-      report = dataElementMergeProcessor.processMerge(params);
+      report = dataElementMergeService.processMerge(params);
     } catch (PersistenceException ex) {
       String helpfulMessage = getHelpfulMessage(ex);
       log.error("Error while processing Data element merge: {}", helpfulMessage);
