+++ dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/controller/indicator/IndicatorController.java
@@ -46,7 +46,7 @@ import org.hisp.dhis.feedback.Status;
 import org.hisp.dhis.i18n.I18nManager;
 import org.hisp.dhis.indicator.Indicator;
 import org.hisp.dhis.merge.MergeParams;
-import org.hisp.dhis.merge.MergeProcessor;
+import org.hisp.dhis.merge.MergeService;
 import org.hisp.dhis.security.RequiresAuthority;
 import org.hisp.dhis.webapi.controller.AbstractCrudController;
 import org.springframework.http.HttpStatus;
@@ -71,7 +71,7 @@ public class IndicatorController extends AbstractCrudController<Indicator> {
 
   private final I18nManager i18nManager;
 
-  private final MergeProcessor indicatorMergeProcessor;
+  private final MergeService indicatorMergeService;
 
   @PostMapping(value = "/expression/description", produces = APPLICATION_JSON_VALUE)
   @ResponseBody
@@ -102,7 +102,7 @@ public class IndicatorController extends AbstractCrudController<Indicator> {
       throws ConflictException {
     log.info("Indicator merge received");
 
-    MergeReport report = indicatorMergeProcessor.processMerge(params);
+    MergeReport report = indicatorMergeService.processMerge(params);
 
     log.info("Indicator merge processed with report: {}", report);
     return WebMessageUtils.mergeReport(report);
